moved quick sort example from recursive package to algorithms
